build(conda): remove noarch
Prior to this commit, the Conda build recipe built the package as
`noarch: python`. However, the `trollius` package is required for
versions of Python prior to 3.4, but must not be included for versions
of Python _since_ 3.4.

In this commit, remove the `noarch` build flag and add a selector to
only require `trollius` on systems with Python versions less than 3.4.